util run_intg_tests: fix running multiple tests in a row

+++ util/run_intg_tests.py
@@ -115,6 +115,7 @@ def wait_backend_ready():
 
     left -= 5
     try:
+        model._core._microscope = None  # force reset of the microscope
         microscope = model.getMicroscope()
         nghosts = len(microscope.ghosts.value) # Components still to start
     except Exception:
@@ -147,6 +148,9 @@ def wait_backend_ready():
 
 def copy_log(log_in_path, log_out_name):
     """ Copy log_in to log_out for later inspection """
+    if not os.path.isfile(log_in_path):
+        logging.info("No logfile %s present", log_in_path)
+        return None
     log_out_path = os.path.join('/tmp', log_out_name)
     logging.debug("Copying log %s to %s", log_in_path, log_out_path)
     shutil.copy(log_in_path, log_out_path)
@@ -180,7 +184,9 @@ def test_config(sim_conf):
     start.start()
 
     # Wait for the back end to load
-    if wait_backend_ready():
+    if not wait_backend_ready():
+        gui = None
+    else:
         logging.info("Starting %s GUI", sim_conf)
         gui = OdemisThread("GUI %s" % sim_conf_fn, CMD_GUI)
         gui.start()
@@ -208,23 +214,24 @@ def test_config(sim_conf):
     try:
         if start.returncode != 0:  # 'ok' return code
             logging.error("Backend failed to start, with return code %d", start.returncode)
-            passed = False
-        elif gui.returncode != 143:  # SIGTERM return code
-            if gui.returncode == 255:
-                logging.warning("Back-end might have not finish loading before the GUI was started")
-            logging.error("GUI failed to start, with return code %d", gui.returncode)
-            passed = False
-        else:
+            return False
+        elif dlog_path:
             # TODO: make error/exception detection in log files more intelligent?
             # TODO: backend always start with an "ERROR" from Pyro, trying to connect to existing backend
             # TODO: differentiate errors happening after asking to stop the back-end
             odemisd_log = open(dlog_path).read()
             for lbl in ERROR_TRIGGER:
                 if lbl in odemisd_log:
-                    logging.error("%s found in back-end log of %s, see %s", lbl, sim_conf_fn, dlog_path)
+                    logging.error("Found %d %s in back-end log of %s, see %s",
+                                  odemisd_log.count(lbl), lbl, sim_conf_fn, dlog_path)
                     passed = False
-                    break
 
+        if gui and gui.returncode != 143:  # SIGTERM return code
+            if gui.returncode == 255:
+                logging.warning("Back-end might have not finish loading before the GUI was started")
+            logging.error("GUI failed to start, with return code %d", gui.returncode)
+            return False
+        elif guilog_path:
             gui_log = open(guilog_path).read()
             for lbl in ERROR_TRIGGER:
                 if lbl in gui_log:
